requirements.txt
fastapi 
uvicorn 
requests 
beautifulsoup4
python-dotenv
python-jose
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart